inicio 22

+++ src/main/java/es/uah/trabajo/juegodelavida/Clases/Elementos/Individuos/Avanzado.java
@@ -2,7 +2,7 @@ package es.uah.trabajo.juegodelavida.Clases.Elementos.Individuos;
 
 public class Avanzado extends Invidiuos {
 
-    public Avanzado(int x, int y, int id, int turnosvida, int probrep, int probclon) {
+    public Avanzado(int x, int y, int id, int turnosvida, float probrep, float probclon) {
         super(x, y, id, turnosvida, probrep, probclon);
     }
 }
