+++ src/main/java/es/uah/trabajo/juegodelavida/Clases/Elementos/Individuos/normal.java
@@ -1,7 +1,7 @@
 package es.uah.trabajo.juegodelavida.Clases.Elementos.Individuos;
 
 public class normal extends Invidiuos{
-    public normal(int x, int y, int id, int turnosvida, int probrep, int probclon) {
+    public normal(int x, int y, int id, int turnosvida, float probrep,float probclon) {
         super(x, y, id, turnosvida, probrep, probclon);
     }
 }
